+++ GetMyTicket.Common/DTOs/Trip/UpdateTripDTO.cs
@@ -14,10 +14,10 @@ namespace GetMyTicket.Common.DTOs.Trip
         public DateTime? EndTime { get; set; }
 
         [Range(0, 9999)]
-        public double? AdultPrice { get; set; }
+        public decimal? AdultPrice { get; set; }
 
         [Required]
         [Range(0, 9999)]
-        public double? ChildrenPrice { get; set; }
+        public decimal? ChildrenPrice { get; set; }
     }
 }
